[graph] Change DGraph.Layers to slice
This change reflects the actual usage of this struct field, i.e. the layers are iterated as an ordered sequence, mostly top-to-bottom. As a matter of fact, this allows to rewrite several for-loops in the form i:=0;i<len(g.Layers);i++ as simple ranges: the index form was used to enforce a deterministic order when iterating over map entries. Indexing the slice with g.Layers[node.Layer] doesn't change the program semantics from when it was a map, as the Layers slice is initialized in postprocessing of phase 2 based on the node.Layer indices.